portfolio_backend/requirements.txt
django>=5.2.4
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
python-dotenv>=1.1.1
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
whitenoise>=6.6.0
redis>=5.0.0